crypto-alert-system/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
google-cloud-pubsub
google-cloud-firestore
requests
python-dotenv
aiohttp